+++ src/render/postprocessing/index.ts
@@ -1,5 +1,6 @@
 export { Bloom, type BloomOptions } from './Bloom';
 export { ColorUber, type ColorUberOptions, type ToneMappingMode } from './ColorUber';
+export { TemporalAA, type TemporalAAOptions } from './TemporalAA';
 export {
     PostProcessRenderPipelineFactory,
     type PostProcessRenderPipelineOptions